impl: Backend Admin Signup
Created the backend API for verifying and eventually executing admin sign up. The verifySignup() function in SignupController calls several other functions to check that the admin username is unique and that the password meets the necessary specifications. If both of those tests pass, then createAdminAccount() is called, which creates the account in the admin table in the database. After failing tests or completing the insertion, a JSON response is returned with status information.

This does not include signup for customers. Since there are separate customer and admin tables, customers will require further development.

GitHub Issue #23