Impl: Backend Customer Signup
Created the backend API for verifying and eventually executing customer sign up. The verifyCustomerSignup() function in SignupController calls several other functions to check that the customer username is unique and that the password meets the necessary specifications. If both of those tests pass, then createCustomerAccount() is called, which creates the account in the customer table in the database. After failing tests or completing the insertion, a JSON response is returned with status information.

Also updated function names from previous commit to distinguish between the customer and admin functions. With this commit, signup for both admins and customers is complete on the backend.

GitHub Issue #23