impl: Backend Add Products
Created the backend API for verifying and eventually adding new products to the products table in the database. The verification first checks that the user is an admin, because non-admins should not be able to add new products. It then checks for duplicate products. A store cannot have multiple products with the same title, though a different store could have a product with the same title. Then it checks that the quantity is at least 1 and if all of those tests have been passed, it adds the product to the database.

GitHub Issue #25