Impl: Backend Admin and Customer Login
Added backend API for customer and admin login. verifyAdminLogin() and verifyCustomerLogin() both call usernameAndPasswordMatch() which calls the appropriate function from the UserAuthenticationDAO to retrieve a user with the same username, then checks that the username and password match.

While implementing the login, I renamed the signupDAO.js file to userAuthenticationDAO.js to better indicate that it is used for both signup and login.

GitHub Issue #23