Impl: Add Password Encryption for Sign Up and Login
Added encryption for passwords that are passed to the backend API for user sign up and login. This means that passwords are encrypted before inserting into the database when a user is signing up. Additionally, when a user is logging in, the enter password is encrypted and then compared to the encrypted password in the database.

Encryption algorithm and key are stored in the .env file so that the key is not public in the repo.

GitHub Issue #23