fix: Fix Bugs in Login and Signup functions
Fix bugs to make login and signup functions work. Modify codes on the backend to make sure the backend can work with the frontend. Delete duplicated functions on the backend. Modify codes on the frontend to make Login connect to the backend. Modify codes on the frontend to make Login and Sign up work together. Delete duplicated functions on the frontend. Delete unused folders and combine folders to make layout clear. Issue #40