impl: Encrypt Passwords in Database
This implements password encryption for user sign up and login. When signing up, the password is encrypted after being passed to the backend, then stored in the database in its encrypted form. When logging in, the password attempt is encrypted and compared to the encrypted password from the database.

This requires the .env file to contain the encryption key and algorithm.

This was previously implemented but was lost in the merge to main.

Issue #44