impl: Basic Add Product Page
This commit implements the add product page but is not styled and image upload is not yet complete. The form on the frontend successfully sends data to the backend, where it is stored in the database. The backend checks for errors and returns error messages that are displayed on the frontend.

The frontend includes the button for uploading the image, but the backend needs further work for actually storing the image.

GitHub Issue #47