impl: Upload Image on Add Product
This commit implements functionality for uploading an image associated with a product when creating a new product on the Add Product page. Images are stored in the backend\uploads folder where the file is updated to contain the name of the product and the adminId for the user. The product entry in the database contains the path and file name for the associated image.

.gitignore has been updated to exclude the individual images inside of the uploads folder, but push the folder itself to the directory.

This currently only allows one image to be associated with a product.

GitHub Issue #56