impl: Product Detail Page
This commit implements the product detail page where a product's details are displayed to the user, as well as the add cart button implemented earlier by Ke.

ProductScreen.js was updated so that the page displays the product's title, image, price, description, category, current stock, and the seller's username.

getProductById() in product.controller.js was updated to return the seller's username with the product.

The uploads folder which contains the uploaded product images was moved to the frontend/public folder to make image display work, so upload.controller.js was updated, as well as the .gitignore and .gitkeep files.

This commit does not complete the entire issue, as we still need to display product images on the homepage to link to the product detail page.

GitHub Issue #60