impl: Backend Route to Return Range of Products
This commit implements the backend route and query necessary to return a number of products from the products tables. The products returned is determined by the page size, currently a constant set equal to 10, and the page number, which is passed as a parameter in the GET request.

The response of this request returns ten products with their details, as well as the total count of products in the table and the number of pages.

GitHub Issue #60