impl: Out of Stock Message on Product Screen
This commit edits the product detail page to display an Out of Stock! message when the quantity is less than 1, instead of showing the number. It still shows the stock quantity if it's greater than 0.

Also in this commit is a minor visual change to the ProductList in the home page, capitalizing Price, adding a dollar sign, and increasing font size of the title.

GitHub Issue #60