fix: Fix Issue with Image Upload
The uploads folder was previously moved from backend to frontend/public so that the images could be accessed and displayed on the frontend. This ended up causing a bug with image uploading, because the image would try to temporarily save in the nonexistent backend/uploads folder before being renamed and moved to the frontend/public/uploads folder.

I resolved this bug by changing the destination in the multer.diskStorage() function, so that the file immediately goes to the frontend folder and then gets renamed there.

I'm including this bug fix here in this branch and under this issue because it's fixing a problem caused by an earlier commit in this branch.

GitHub Issue #60