impl: Out of Stock Alert
This commit adds an alert window to notify user that the item is out of stock and cannot be added to the cart. A previous commit changed the quantity to show Out of Stock instead of the quantity number, but didn't notify user after clicking the Add to Cart button. Now, an alert window pops up in the browser to notify user more visibly.

GitHub Issue #60