fix: Add Key to Products
The developer consoler in browser was showing "Warning: Each child in a list should have a unique "key" prop.". This has been resolved by adding key={product.title} to the outermost div inside of the products mapping.

Ke originally recommend placing this key value inside of the Image element, but I found that this did not resolve the warning, while putting it in the outermost div does.

GitHub Issue #60