impl: Store Manager Order History Page
A previous commit implemented the backend API for fetching orders and returning the info to the frontend. This commit implements the frontend necessary to process and display the store's order history info page.

-App.js was updated to add the route to the page.

-header.js was updated to add a dropdown option to the Store Manager dropdown menu that links to the page.

- StoreOrdersScreen was created to display the store's order history.

The page will display a message if the user is not logged in as a store manager or if the store has no orders to display. If a store has orders it shows a table with order summaries and then more detailed order info below the table.

GitHub Issue #73