refactor: Combine Code to Process Store's Order
A store requires an order from the database to be reprocessed in two regards before being displayed on the store manager's order history page.

1. If an order has items from multiple stores, the items in the order that do not belong to the store should be removed from the order so that a store manager only sees the items ordered from their store.
2. The total price of the order has to be recalculated to include only the price of that store's items, rather than the price of the whole order.

The first aspect was originally handled in the backend before returning to the frontend, and the second aspect was originally handled in the frontend. This created some duplicate code because both required nested for loops to go through each order and each item in each order.

I reduced some of the repeat code by combining both parts into the backend.

GitHub Issue #73