Fix: Remove Unique Key Prop Warning
The console was displaying a "Warning: Each child in a list should have a unique "key" prop." warning. Resolving this required adding key={order._id} or key={item._id} within the map functions. This assigned unique keys to each child of each list.

The shorthand <> had to be updated to the explicit <React.Fragment> in order to assign a key value.

GitHub Issue #73