Fix: Prevent Errors After Log Out
If the user is logged in and accesses the admin product list screen or the product edit screen, then logouts while on that page, the site will show errors.

This commit modifies AdminProductListScreen.js and ProductEditScreen.js and checks to make sure that the user is logged in as an admin. If the user is not logged in as an admin, then a message is shown on the page and no errors are generated.

GitHub Issue #71